+++ packages/@expo/cli/src/run/ensureNativeProject.ts
@@ -8,7 +8,7 @@ import { profile } from '../utils/profile';
 
 export async function ensureNativeProjectAsync(
   projectRoot: string,
-  { platform, install }: { platform: ModPlatform; install?: boolean }
+  { platform, install, template }: { platform: ModPlatform; install?: boolean, template?: string }
 ) {
   // If the user has an empty android folder then the project won't build, this can happen when they delete the prebuild files in git.
   // Check to ensure most of the core files are in place, and prompt to remove the folder if they aren't.
@@ -19,6 +19,7 @@ export async function ensureNativeProjectAsync(
     await prebuildAsync(projectRoot, {
       install: !!install,
       platforms: [platform],
+      template,
     });
   } else {
     return true;